backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic>=2.7.4,<3.0.0
requests==2.31.0
pandas==2.2.0
langchain>=0.1.4
langchain-core>=0.1.16
langchain-openai>=0.0.5
langgraph>=0.0.26
openai==1.59.7
python-dateutil==2.9.0.post0
numpy==1.26.4
matplotlib==3.10.0
dnspython==2.5.0 